Time: 0 ms (100.00%) | Memory: 7.9 MB (26.11%) - LeetSync

+++ 118-pascals-triangle/pascals-triangle.cpp
@@ -5,7 +5,7 @@ public:
 //     return n*fact(n-1);
 // }
 unsigned long long f(int n,int r){
-    int res=1;
+    unsigned long long res=1;
     for(int i=0;i<r;i++){
         res*=(n-i);
         res/=(i+1);
